+++ src/api/cart/createCart.ts
@@ -1,26 +1,8 @@
-import type { ByProjectKeyRequestBuilder, Cart, ClientResponse } from '@commercetools/platform-sdk';
-import { getAnonymousFlowClient, getRefreshFlowClient } from '../BuildClient';
-import { apiRoot } from '../apiRoot';
-
-async function getCustomerFlow(): Promise<ByProjectKeyRequestBuilder> {
-  const refreshToken = localStorage.getItem('sloth-refreshToken');
-  let customerFlow;
-
-  if (refreshToken) {
-    try {
-      customerFlow = getRefreshFlowClient();
-    } catch {
-      customerFlow = getAnonymousFlowClient();
-    }
-  } else {
-    customerFlow = getAnonymousFlowClient();
-  }
-
-  return customerFlow;
-}
+import type { Cart } from '@commercetools/platform-sdk';
+import { getClientFlow } from '../BuildClient';
 
 export const createCart = async (): Promise<Cart> => {
-  const client = await getCustomerFlow();
+  const client = await getClientFlow();
   const response = await client
     .me()
     .carts()
@@ -31,28 +13,4 @@ export const createCart = async (): Promise<Cart> => {
     })
     .execute();
   return response.body;
-};
-
-export const addItemToCart = async (cartId: string, productId: string, version: number): Promise<Cart> => {
-  const response = await apiRoot
-    .carts()
-    .withId({ ID: cartId })
-    .post({
-      body: {
-        version,
-        actions: [
-          {
-            action: 'addLineItem',
-            productId,
-            quantity: 1
-          }
-        ]
-      }
-    })
-    .execute();
-  return response.body;
-};
-
-export const getCart = async (customerId: string): Promise<ClientResponse<Cart>> => {
-  return await apiRoot.carts().withCustomerId({ customerId }).get().execute();
 };
