+++ src/api/cart/getCart.ts
@@ -1,0 +1,6 @@
+import type { ClientResponse, Cart } from '@commercetools/platform-sdk';
+import { apiRoot } from '../apiRoot';
+
+export const getCart = async (customerId: string): Promise<ClientResponse<Cart>> => {
+  return await apiRoot.carts().withCustomerId({ customerId }).get().execute();
+};
